pyproject.toml
[project]
name = "rtt"
version = "0.1.0"
description = "Semantic video search engine over public domain films"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "lancedb>=0.20",
    "pyarrow>=18",
    "python-dotenv>=1.0",
    "anthropic>=0.43",
    "faster-whisper>=1.1",
    "httpx>=0.28",
]

[project.scripts]
rtt = "rtt.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rtt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "pytest-recording>=0.13",
    "vcrpy>=7",
    "httpx>=0.28",
]
